small refactor + IndexedGeoTiffLayer logging
