+++ spark/src/main/scala/geotrellis/spark/io/geotiff/SinglebandGeoTiffRDDLayerReader.scala
@@ -2,7 +2,7 @@ package geotrellis.spark.io.geotiff
 
 import geotrellis.proj4.WebMercator
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader
-import geotrellis.raster.{CellGrid, Raster, Tile}
+import geotrellis.raster.{Raster, Tile}
 import geotrellis.spark.io.hadoop.{HdfsRangeReader, HdfsUtils}
 import geotrellis.spark.tiling.ZoomedLayoutScheme
 import geotrellis.spark.util._
@@ -15,14 +15,6 @@ import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 
 import java.net.URI
-
-trait GeoTiffRDDLayerReader[T <: CellGrid] {
-  val rdd: RDD[(Extent, URI)]
-  val discriminator: URI => String
-
-  def read(layerId: LayerId)(x: Int, y: Int)(implicit sc: SparkContext): Raster[T]
-  def readAll(layerId: LayerId)(implicit sc: SparkContext): RDD[Raster[T]]
-}
 
 case class SinglebandGeoTiffRDDLayerReader(
   rdd: RDD[(Extent, URI)],
